+++ src/world/coordinates.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Coordinates {
+    pub x: i32,
+    pub y: i32,
+}
