+++ src/world/npc.rs
@@ -1,0 +1,30 @@
+use serde::{Deserialize, Serialize};
+use super::coordinates::Coordinates;
+use super::entity::EntityType;
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct NPC {
+    pub type_: EntityType,
+    pub name: String,
+    pub position: Coordinates,
+    pub home: Coordinates,
+    pub variation_index: i32,
+}
+
+impl NPC {
+    pub fn new(
+        type_: EntityType,
+        name: String,
+        position: Coordinates,
+        home: Coordinates,
+        variation_index: i32,
+    ) -> Self {
+        Self {
+            type_,
+            name,
+            position,
+            home,
+            variation_index,
+        }
+    }
+}
