+++ src/world/sign.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+use super::coordinates::Coordinates;
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Sign {
+    pub text: String,
+    pub position: Coordinates,
+}
